fix(consumer-tests): Further improve the error message when a consumer test doesn't seem to have the mock server running

+++ src/httpPact.ts
@@ -12,6 +12,16 @@ import { Server } from "@pact-foundation/pact-node/src/server"
 import VerificationError from "./errors/verificationError"
 import ConfigurationError from "./errors/configurationError"
 
+const logErrorNoMockServer = () => {
+  logger.error(
+    "The pact mock service doesn't appear to be running\n" +
+      "  - Please check the logs above to ensure that there are no pact service startup failures\n" +
+      "  - Please check that pact lifecycle methods are called in the correct order (setup() needs to be called before this method)\n" +
+      "  - Please check that your test code waits for the promises returned from lifecycle methods to complete before calling the next one\n" +
+      "  - To learn more about what is happening during your pact run, try setting logLevel: 'DEBUG'"
+  )
+}
+
 /**
  * Creates a new {@link PactProvider}.
  * @memberof Pact
@@ -89,6 +99,14 @@ export class Pact {
   public addInteraction(
     interactionObj: InteractionObject | Interaction
   ): Promise<string> {
+    if (!this.mockService) {
+      logErrorNoMockServer()
+      return Promise.reject(
+        new Error(
+          "The pact mock service wasn't running when addInteraction was called"
+        )
+      )
+    }
     if (interactionObj instanceof Interaction) {
       return this.mockService.addInteraction(interactionObj)
     }
@@ -113,13 +131,11 @@ export class Pact {
    */
   public verify(): Promise<string> {
     if (!this.mockService) {
+      logErrorNoMockServer()
       return Promise.reject(
-        new Error(
-          "The Pact mock service does not appear to be running. Please check the log above, or try setting `logLevel: 'DEBUG'`"
-        )
+        new Error("The pact mock service wasn't running when verify was called")
       )
     }
-
     return this.mockService
       .verify()
       .then(() => this.mockService.removeInteractions())
@@ -156,9 +172,10 @@ export class Pact {
     this.finalized = true
 
     if (!this.mockService) {
+      logErrorNoMockServer()
       return Promise.reject(
         new Error(
-          "The Pact mock service does not appear to be running. Please check the log above, or try setting `logLevel: 'DEBUG'`"
+          "The pact mock service wasn't running when finalize was called"
         )
       )
     }
@@ -197,6 +214,14 @@ export class Pact {
    * @returns {Promise}
    */
   public writePact(): Promise<string> {
+    if (!this.mockService) {
+      logErrorNoMockServer()
+      return Promise.reject(
+        new Error(
+          "The pact mock service wasn't running when writePact was called"
+        )
+      )
+    }
     return this.mockService.writePact()
   }
 
@@ -207,6 +232,14 @@ export class Pact {
    * @returns {Promise}
    */
   public removeInteractions(): Promise<string> {
+    if (!this.mockService) {
+      logErrorNoMockServer()
+      return Promise.reject(
+        new Error(
+          "The pact mock service wasn't running when removeInteractions was called"
+        )
+      )
+    }
     return this.mockService.removeInteractions()
   }
